Move field CRDTs to core.crdt
