+++ internal/core/crdt/field.go
@@ -8,7 +8,7 @@
 // by the Apache License, Version 2.0, included in the file
 // licenses/APL.txt.
 
-package merklecrdt
+package crdt
 
 import (
 	"github.com/sourcenetwork/defradb/client"
